pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragpipe"
version = "0.1.0"
description = "Declarative RAG data pipeline library. Sources → Transforms → Sinks for vector databases."
readme = "README.md"
license = "BUSL-1.1"
license-files = { paths = ["LICENSE"] }
requires-python = ">=3.10"
authors = [
    { name = "avasis-ai", email = "avasis@proton.me" },
]
keywords = ["rag", "pipeline", "vector-database", "embeddings", "llm", "data-ingestion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Business Source License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.25",
]

[project.optional-dependencies]
web = ["beautifulsoup4>=4.12", "httpx>=0.25"]
qdrant = ["qdrant-client>=1.7"]
pinecone = ["pinecone-client>=3.0"]
embeddings = ["openai>=1.0"]
all = [
    "ragpipe[web]",
    "ragpipe[qdrant]",
    "ragpipe[pinecone]",
    "ragpipe[embeddings]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/avasis-ai/ragpipe"
Repository = "https://github.com/avasis-ai/ragpipe"
Issues = "https://github.com/avasis-ai/ragpipe/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ragpipe"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
